requirements.txt
pyqt6==6.0.2
pyqt6-tools==6.0.2.3.2
pillow
pyinstaller